consolidate libs, ensure markets are refreshed during pool operations and add e2e test to cover for potentially stale market positions
