+++ .prettierignore
@@ -9,3 +9,4 @@ package-lock.json
 README.md
 types
 keypairs
+packages/margin/src/fixed-term/types/JetMarket.ts
