+++ apps/react-app/cypress/support/actions.ts
@@ -44,8 +44,9 @@ export const airdrop = (symbol: string, asset: string) => {
 export const deposit = (symbol: string, amount: number) => {
   cy.get(`.${symbol}-pools-table-row`).click();
   cy.contains('button', 'Deposit').should('not.be.disabled').click();
-  const input = cy.get('.ant-modal-content input.ant-input').should('not.be.disabled');
-  input.type(`${amount}`);
+  cy.get('.ant-modal-content input.ant-input').as('input')
+  cy.get('@input').should('not.be.disabled');
+  cy.get('@input').type(`${amount}`);
   cy.get('.ant-modal-body button.ant-btn').should('not.be.disabled').contains('Deposit').click();
   cy.contains(`Your deposit of ${formatWithCommas(amount)} ${symbol} was successfully processed.`);
 };
@@ -57,17 +58,24 @@ export const borrow = (symbol: string, amount: number, resetMaxState?: boolean) 
     // Reset max trade values to simulate borrowing on existing account
     cy.get('[data-testid="reset-max-trade"]').click();
   }
-  const input = cy.get('.ant-modal-content input.ant-input').should('not.be.disabled');
-  input.click().type(`${amount}`);
-  cy.get('.ant-modal-body button.ant-btn').should('not.be.disabled').contains('Borrow').click();
+  cy.wait(500)
+  cy.get('.ant-modal-content input.ant-input').as('input')
+  cy.get('@input').should('not.be.disabled');
+  cy.get('@input').click()
+  cy.get('@input').type(`${amount}`);
+  cy.get('.ant-modal-body button.ant-btn.ant-btn-default.ant-btn-block').as('submitButton')
+  cy.get('@submitButton').should('not.be.disabled')
+  cy.get('@submitButton').click();
   cy.contains('borrow successful');
 };
 
 export const withdraw = (symbol: string, amount: number) => {
   cy.get(`.${symbol}-pools-table-row`).click();
   cy.get(`.account-snapshot-footer button`).contains('Withdraw').click();
-  const input = cy.get('.ant-modal-content input.ant-input').should('not.be.disabled');
-  input.click().type(`${amount}`);
+  cy.get('.ant-modal-content input.ant-input').as('input')
+  cy.get('@input').should('not.be.disabled');
+  cy.get('@input').click()
+  cy.get('@input').type(`${amount}`);
   cy.get('.ant-modal-body button.ant-btn').should('not.be.disabled').contains('Withdraw').click();
   cy.contains('withdraw successful');
 };
@@ -82,8 +90,10 @@ export const repay = (symbol: string, amount: number, fromDeposit: boolean) => {
   cy.contains('Repay From Wallet')
     .siblings()
     .should(fromDeposit ? 'not.have.class' : 'have.class', 'ant-switch-checked');
-  const input = cy.get('.ant-modal-content input.ant-input').should('not.be.disabled');
-  input.click().type(`${amount}`);
+  cy.get('.ant-modal-content input.ant-input').as('input')
+  cy.get('@input').should('not.be.disabled');
+  cy.get('@input').click()
+  cy.get('@input').type(`${amount}`);
   cy.get('.ant-modal-body button.ant-btn').should('not.be.disabled').contains('Repay').click();
   cy.contains(`${symbol} was successfully processed.`);
 };
