+++ apps/react-app/src/components/fixed-term/borrow-entry/request-loan.tsx
@@ -1,6 +1,6 @@
 import { Button, InputNumber, Switch, Tooltip } from 'antd';
 import { formatDuration, intervalToDuration } from 'date-fns';
-import { MarketAndconfig, requestLoan } from '@jet-lab/fixed-term';
+import { MarketAndconfig, requestLoan } from '@jet-lab/margin';
 import { notify } from '@utils/notify';
 import { getExplorerUrl } from '@utils/ui';
 import BN from 'bn.js';
