+++ apps/react-app/src/components/fixed-term/lend-entry/offer-loan.tsx
@@ -1,6 +1,6 @@
 import { Button, InputNumber, Switch, Tooltip } from 'antd';
 import { formatDuration, intervalToDuration } from 'date-fns';
-import { MarketAndconfig, offerLoan } from '@jet-lab/fixed-term';
+import { MarketAndconfig, offerLoan } from '@jet-lab/margin';
 import { notify } from '@utils/notify';
 import { getExplorerUrl } from '@utils/ui';
 import BN from 'bn.js';
