+++ apps/react-app/src/components/fixed-term/shared/fixed-term-market-chart.tsx
@@ -15,7 +15,7 @@ import { friendlyMarketName } from '@utils/jet/fixed-term-utils';
 import { useMemo } from 'react';
 import { calculate_implied_price, price_to_rate } from '@jet-lab/wasm';
 import { MainConfig } from '@state/config/marginConfig';
-import { MarketAndconfig } from '@jet-lab/fixed-term';
+import { MarketAndconfig } from '@jet-lab/margin';
 interface FixedChart {
   type: 'bids' | 'asks';
 }
