+++ apps/react-app/src/utils/jet/marginActions.ts
@@ -13,6 +13,7 @@ import { useProvider } from './provider';
 import { NOTIFICATION_DURATION } from '../notify';
 import { message } from 'antd';
 import { Cluster } from '@state/settings/settings';
+import { AllFixedTermMarketsAtom } from '@state/fixed-term/fixed-term-market-sync';
 
 export enum ActionResponse {
   Success = 'SUCCESS',
@@ -26,6 +27,7 @@ export function useMarginActions() {
   const dictionary = useRecoilValue(Dictionary);
   const { programs, provider } = useProvider();
   const pools = useRecoilValue(Pools);
+  const markets = useRecoilValue(AllFixedTermMarketsAtom)
   const currentPool = useRecoilValue(CurrentPool);
   const wallet = useWallet();
   const walletTokens = useRecoilValue(WalletTokens);
@@ -178,6 +180,7 @@ export function useMarginActions() {
       const txId = await currentPool.withdraw({
         marginAccount: currentAccount,
         pools: Object.values(pools.tokenPools),
+        markets: markets.map(m => m.market),
         destination: token.address,
         change
       });
@@ -195,15 +198,17 @@ export function useMarginActions() {
 
   // Borrow
   async function borrow(): Promise<[string | undefined, ActionResponse]> {
-    if (!pools || !currentPool || !walletTokens || !currentAccount || !accountPoolPosition) {
+    if (!pools || !currentPool || !walletTokens || !currentAccount || !accountPoolPosition  || !markets) {
       console.error('Accounts and/or pools not loaded');
       throw new Error();
     }
 
     try {
+      console.log(markets)
       const txId = await currentPool.marginBorrow({
         marginAccount: currentAccount,
         pools: Object.values(pools.tokenPools),
+        markets: markets.map(m => m.market),
         change: PoolTokenChange.shiftBy(tokenInputAmount)
       });
       await actionRefresh();
@@ -235,6 +240,7 @@ export function useMarginActions() {
         marginAccount: currentAccount,
         source: accountRepay ? undefined : token.address,
         pools: Object.values(pools.tokenPools),
+        markets: markets.map(m => m.market),
         change,
         closeLoan
       });
@@ -268,6 +274,7 @@ export function useMarginActions() {
       const txId = await inputToken.splTokenSwap({
         marginAccount: currentAccount,
         pools: Object.values(pools.tokenPools),
+        markets: markets.map(m => m.market),
         outputToken,
         swapPool,
         swapAmount,
@@ -295,7 +302,7 @@ export function useMarginActions() {
   // Transfer
   async function transfer(
     fromAccount: MarginAccount,
-    toAccount: MarginAccount
+    toAccount: MarginAccount,
   ): Promise<[string | undefined, ActionResponse]> {
     if (!pools || !currentPool || !currentAccount || !fromAccount.walletTokens || !toAccount.walletTokens) {
       console.error('Accounts and/or pools not loaded');
@@ -312,15 +319,15 @@ export function useMarginActions() {
     );
     try {
       // Refresh positions
-      await currentPool.withPrioritisedPositionRefresh({
+      await fromAccount.withPrioritisedPositionRefresh({
         instructions: refreshInstructions,
         pools: pools.tokenPools,
-        marginAccount: fromAccount
+        markets: markets.map(m => m.market),
       });
-      await currentPool.withPrioritisedPositionRefresh({
+      await toAccount.withPrioritisedPositionRefresh({
         instructions: refreshInstructions,
         pools: pools.tokenPools,
-        marginAccount: toAccount
+        markets: markets.map(m => m.market),
       });
 
       // toAccount deposit position
