--- packages/fixed-term/.prettierrc
@@ -1,10 +1,0 @@
-{
-  "arrowParens": "avoid",
-  "bracketSameLine": true,
-  "bracketSpacing": true,
-  "endOfLine": "lf",
-  "printWidth": 120,
-  "semi": false,
-  "singleQuote": false,
-  "trailingComma": "none"
-}
