--- packages/fixed-term/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@jet-lab/fixed-term",
-  "version": "0.1.0",
-  "description": "Library for interacting with the Jet fixed term market on-chain programs",
-  "type": "module",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "src/*",
-    "dist/*"
-  ],
-  "license": "AGPL-3.0-or-later",
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "clean": "rm -rf ./dist & rm -rf .turbo & rm -f tsconfig.tsbuildinfo",
-    "lint": "tslint -c tslint.json src/**/*.ts",
-    "prepublishOnly": "yarn build",
-    "watch": "tsc --watch",
-    "fmt": "prettier src/ --check",
-    "fmt:fix": "prettier src/ --write"
-  },
-  "dependencies": {
-    "@jet-lab/margin": "*",
-    "@jet-lab/wasm": "*",
-    "@project-serum/anchor": "0.24.2",
-    "@solana/spl-token": "0.3.6",
-    "@solana/web3.js": "1.66.2"
-  },
-  "devDependencies": {
-    "prettier": "^2.3.2",
-    "tslint": "^6.1.3",
-    "typescript": "^4.9.3"
-  }
-}
