+++ packages/margin/package.json
@@ -43,6 +43,7 @@
   "dependencies": {
     "@project-serum/anchor": "0.24.2",
     "@pythnetwork/client": "^2.7.0",
+    "@jet-lab/wasm": "*",
     "@solana/buffer-layout": "^4.0.0",
     "@solana/spl-token": "0.3.6",
     "@solana/web3.js": "1.66.2",
