+++ packages/margin/src/index.ts
@@ -19,3 +19,4 @@ export * from "./margin"
 export * from "./token"
 export * from "./types"
 export * from "./utils"
+export * from './fixed-term'
